FIX: test_via_design.py
